+++ src/Scaffolding/VS.Web.CG.Core/CodeGeneratorActionsService.cs
@@ -71,7 +71,7 @@ namespace Microsoft.VisualStudio.Web.CodeGeneration
 
             Debug.Assert(_fileSystem.FileExists(templatePath));
             var templateContent = _fileSystem.ReadAllText(templatePath);
-
+            
             var templateResult = await _templatingService.RunTemplateAsync(templateContent, templateModel);
 
             if (templateResult.ProcessingException != null)
