+++ src/Scaffolding/VS.Web.CG.Mvc/Templates/Identity/Bootstrap4/Pages/Account/Account.ConfirmEmail.cs.cshtml
@@ -17,7 +17,7 @@ using Microsoft.AspNetCore.Authorization;
             "Microsoft.AspNetCore.WebUtilities",
         });
 
-    var thisNamespace = $"{Model.Namespace}.Areas.Identity.Pages.Account";
+    var thisNamespace = @Model.Namespace + ".Areas.Identity.Pages.Account";
     if (!string.IsNullOrEmpty(Model.UserClassNamespace) && thisNamespace != Model.UserClassNamespace)
     {
         namespaceSet.Add(Model.UserClassNamespace);
