+++ src/Scaffolding/VS.Web.CG.Mvc/Templates/Identity/Bootstrap4/Pages/Account/Account.ExternalLogin.cs.cshtml
@@ -22,7 +22,7 @@ using Microsoft.AspNetCore.Authorization;
             "Microsoft.Extensions.Logging",
         });
 
-    var thisNamespace = $"{Model.Namespace}.Areas.Identity.Pages.Account";
+    var thisNamespace =@Model.Namespace + ".Areas.Identity.Pages.Account";
     if (!string.IsNullOrEmpty(Model.UserClassNamespace) && thisNamespace != Model.UserClassNamespace)
     {
         namespaceSet.Add(Model.UserClassNamespace);
