+++ src/Scaffolding/VS.Web.CG.Mvc/Templates/Identity/Bootstrap4/Pages/Account/Account.RegisterConfirmation.cs.cshtml
@@ -15,7 +15,7 @@ using System.Threading.Tasks;
             "Microsoft.AspNetCore.WebUtilities",
         });
 
-    var thisNamespace = $"{Model.Namespace}.Areas.Identity.Pages.Account";
+    var thisNamespace = @Model.Namespace + ".Areas.Identity.Pages.Account";
     if (!string.IsNullOrEmpty(Model.UserClassNamespace) && thisNamespace != Model.UserClassNamespace)
     {
         namespaceSet.Add(Model.UserClassNamespace);
