+++ src/Scaffolding/VS.Web.CG.Mvc/Templates/Identity/Bootstrap4/Pages/Account/Manage/Account.Manage.DeletePersonalData.cs.cshtml
@@ -13,7 +13,7 @@ using System.Threading.Tasks;
             "Microsoft.AspNetCore.Identity",
             "Microsoft.Extensions.Logging",
         });
-    var thisNamespace = $"{Model.Namespace}.Areas.Identity.Pages.Account.Manage";
+    var thisNamespace = @Model.Namespace + ".Areas.Identity.Pages.Account.Manage";
     if (!string.IsNullOrEmpty(Model.UserClassNamespace) && thisNamespace != Model.UserClassNamespace)
     {
         namespaceSet.Add(Model.UserClassNamespace);
