+++ src/Scaffolding/VS.Web.CG.Mvc/Templates/Identity/Bootstrap4/Pages/Account/Manage/Account.Manage.TwoFactorAuthentication.cs.cshtml
@@ -15,7 +15,7 @@ using System.Threading.Tasks;
             "Microsoft.Extensions.Logging",
         });
 
-    var thisNamespace = $"{Model.Namespace}.Areas.Identity.Pages.Account.Manage";
+    var thisNamespace = @Model.Namespace + ".Areas.Identity.Pages.Account.Manage";
     if (!string.IsNullOrEmpty(Model.UserClassNamespace) && thisNamespace != Model.UserClassNamespace)
     {
         namespaceSet.Add(Model.UserClassNamespace);
