+++ src/Shared/Microsoft.DotNet.Scaffolding.Shared/CodeModifier/CodeChange/CodeFile.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Microsoft.DotNet.MSIdentity.CodeReaderWriter.CodeChange
+namespace Microsoft.DotNet.Scaffolding.Shared.CodeModifier.CodeChange
 {
     public class CodeFile
     {
