+++ src/Shared/Microsoft.DotNet.Scaffolding.Shared/IConsoleLogger.cs
@@ -2,12 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.DotNet.MSIdentity.Tool
+namespace Microsoft.DotNet.Scaffolding.Shared
 {
     public interface IConsoleLogger
     {
-        void LogMessage(string? message, LogMessageType level, bool removeNewLine = false);
-        void LogMessage(string? message, bool removeNewLine = false);
+        void LogMessage(string message, LogMessageType level, bool removeNewLine = false);
+        void LogMessage(string message, bool removeNewLine = false);
         void LogJsonMessage(JsonResponse jsonMessage);
     }
